requirements/base.txt
django
psycopg2-binary
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-filter
django-modeltranslation
django-redis
cryptography==41.0.7
django-recaptcha==3.0.0
nplusone==1.0.0
celery==5.3.6
redis==5.0.1
flower==2.0.1
dj-database-url==2.2.0
phonenumbers==9.0.7
phonenumberslite==9.0.7
django-phonenumber-field==8.1.0
djangorestframework_simplejwt==5.5.0
requests==2.32.4


